renames point_earning file
