+++ src/bbc_game/configs/game_config.py
@@ -1,10 +1,10 @@
 from bbc_game._typing import ShopType
 from typing import Optional
-from bbc_game.point_earning import PointDistributer, PointDistributerFactory
+from bbc_game.point_distribution import PointDistributer, PointDistributerFactory
 
 class GameConfig:
-    def __init__(self, point_earning: PointDistributer, shop: Optional[ShopType] = None, base_currency: float = 0, base_earn_rate: float = 0,
-                 base_modifier: float = 1):
+    def __init__(self, point_earning: PointDistributer, shop: Optional[ShopType] = None, base_currency: float = 0,
+                 base_earn_rate: float = 0, base_modifier: float = 1):
         self.point_earning = point_earning
         self.shop = shop
         self.base_currency = base_currency
@@ -12,8 +12,8 @@ class GameConfig:
         self.base_modifier = base_modifier
 
 class GameConfigFactory:
-    def __init__(self, point_earning_factory: PointDistributerFactory, shop: Optional[ShopType] = None, base_currency: float = 0, base_earn_rate: float = 0,
-                 base_modifier: float = 1):
+    def __init__(self, point_earning_factory: PointDistributerFactory, shop: Optional[ShopType] = None,
+                 base_currency: float = 0, base_earn_rate: float = 0, base_modifier: float = 1):
         self.shop = shop
         self.base_currency = base_currency
         self.base_earn_rate = base_earn_rate
