+++ src/bbc_game/point_distribution.py
